Moved include file. Added libpcan dependency.
