+++ allegro_hand_common/package.xml
@@ -21,8 +21,10 @@
   <!--   <test_depend>gtest</test_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>roscpp</build_depend>
+  <build_depend>libpcan</build_depend>
   <build_depend>std_msgs</build_depend>
   <run_depend>roscpp</run_depend>
+  <run_depend>libpcan</run_depend>
   <run_depend>std_msgs</run_depend>
   <run_depend>message_runtime</run_depend>
